refactor : 예외 관련 공통 클래스 패키지 분리 작업
