--- src/main/java/roomescape/time/global/error/exception/InvalidValueException.java
@@ -1,13 +1,0 @@
-package roomescape.time.global.error.exception;
-
-public class InvalidValueException extends UserException {
-
-    public InvalidValueException(ErrorCode errorCode) {
-        super(errorCode.getStatus(), errorCode.getErrorMessage());
-    }
-
-    @Override
-    public int getStatus() {
-        return super.getStatus();
-    }
-}
